+++ native/src/installation_reader.cpp
@@ -661,6 +661,7 @@ void InstallationReader::launch_impl(Dart_Port port, const char* app_id, const c
 
     FpInstance info;
     info.appId = safe_str(flatpak_instance_get_app(instance));
+    info.runtime = safe_str(flatpak_instance_get_runtime(instance));
     info.instanceId = safe_str(flatpak_instance_get_id(instance));
     info.arch = safe_str(flatpak_instance_get_arch(instance));
     info.branch = safe_str(flatpak_instance_get_branch(instance));
@@ -894,6 +895,7 @@ void InstallationReader::list_running(Dart_Port port) {
         auto* inst = static_cast<FlatpakInstance*>(instances->pdata[i]);
         FpInstance info;
         info.appId = safe_str(flatpak_instance_get_app(inst));
+        info.runtime = safe_str(flatpak_instance_get_runtime(inst));
         info.instanceId = safe_str(flatpak_instance_get_id(inst));
         info.arch = safe_str(flatpak_instance_get_arch(inst));
         info.branch = safe_str(flatpak_instance_get_branch(inst));